+++ tests/testthat/test-lifecycle-deprecated.R
@@ -12,3 +12,11 @@ test_that("vec_duplicate_id() still works", {
   local_lifecycle_silence()
   expect_identical(vec_duplicate_id(c(1, 1, 3)), vec_first_loc(c(1, 1, 3)))
 })
+
+test_that("vec_duplicate_detect() still works", {
+  local_lifecycle_silence()
+  expect_identical(
+    vec_duplicate_detect(c(1, 1, 3)),
+    vec_duplicate_flg(c(1, 1, 3), first = TRUE)
+  )
+})